hw05-jdbc-library - испрваление замечаний

+++ hw05-jdbc-library/src/main/resources/schema.sql
@@ -19,8 +19,8 @@ CREATE TABLE books
     id        BIGINT AUTO_INCREMENT,
     title     VARCHAR(255) NOT NULL,
     year      INT,
-    author_id INT,
-    genre_id  INT,
+    author_id BIGINT,
+    genre_id  BIGINT,
     PRIMARY KEY (id),
     FOREIGN KEY (author_id) REFERENCES authors (id) ON DELETE SET NULL,
     FOREIGN KEY (genre_id) REFERENCES genres (id) ON DELETE SET NULL
